Add landing page while maintaining react state
