--- src/app/components/LandingPage.tsx
@@ -1,9 +1,0 @@
-const LandingPage = () => {
-  return (
-    <div>
-      <h1>Welcome to the Doula Medicaid Application</h1>
-    </div>
-  );
-};
-
-export default LandingPage;
