+++ src/app/form/clientRoutes.tsx
@@ -1,4 +1,3 @@
-import LandingPage from "@/app/components/LandingPage";
 import BusinessDetailsStep1 from "@/app/form/(formSteps)/business-details/1/BusinessDetailsStep1";
 import BusinessDetailsStep2 from "@/app/form/(formSteps)/business-details/2/BusinessDetailsStep2";
 import BusinessDetailsStep3 from "@/app/form/(formSteps)/business-details/3/BusinessDetailsStep3";
@@ -7,6 +6,7 @@ import FinishSection from "@/app/form/(formSteps)/finish/FinishSection";
 import { FormLayout } from "@/app/form/(formSteps)/FormLayout";
 import InsuranceStep1 from "@/app/form/(formSteps)/insurance/1/InsuranceStep1";
 import InsuranceStep2 from "@/app/form/(formSteps)/insurance/2/InsuranceStep2";
+import LandingPage from "@/app/form/(formSteps)/landing/LandingPage";
 import PersonalDetailsStep1 from "@/app/form/(formSteps)/personal-details/1/PersonalDetailsStep1";
 import PersonalDetailsStep2 from "@/app/form/(formSteps)/personal-details/2/PersonalDetailsStep2";
 import PersonalDetailsStep3 from "@/app/form/(formSteps)/personal-details/3/PersonalDetailsStep3";
@@ -18,9 +18,9 @@ import { BrowserRouter, Route, Routes } from "react-router";
 
 export const routes = (
   <Routes>
-    <Route index element={<LandingPage />}></Route>
     <Route path="form">
       <Route element={<FormLayout />}>
+        <Route path="landing" element={<LandingPage />} />
         <Route path="screening">
           <Route path="1" element={<ScreeningStep1 />} />
           <Route path="2" element={<ScreeningStep2 />} />
