+++ src/app/form/components/ProgressBar.tsx
@@ -1,0 +1,83 @@
+import { getCurrentFormProgress, progressBarSections } from "@form/_utils/formProgress";
+import { RequiredMarker } from "@trussworks/react-uswds";
+import { usePathname } from "next/navigation";
+
+type CompletionState = "complete" | "current" | "incomplete";
+
+export const ProgressBar = () => {
+  const pathname = usePathname();
+  const { section: currentSection, step: currentStep } = getCurrentFormProgress(pathname);
+  const currentSectionIndex = progressBarSections.findIndex(
+    (sections) => sections.id === currentSection.id,
+  );
+
+  return (
+    <div className="usa-step-indicator" aria-label="progress">
+      <ol className="usa-step-indicator__segments">
+        {progressBarSections.map((section, sectionIndex) => {
+          let completionState: CompletionState;
+          switch (true) {
+            case sectionIndex < currentSectionIndex:
+              completionState = "complete";
+              break;
+            case sectionIndex === currentSectionIndex:
+              completionState = "current";
+              break;
+            case sectionIndex > currentSectionIndex:
+              completionState = "incomplete";
+              break;
+            default:
+              throw new Error(`Unexpected logic path: ${sectionIndex}, ${currentSectionIndex}`);
+              break;
+          }
+
+          const liSegmentClassSuffix = {
+            complete: "complete",
+            current: "current",
+            incomplete: null,
+          }[completionState];
+          const screenreaderStatus = {
+            complete: "completed",
+            current: null,
+            incomplete: "not completed",
+          }[completionState];
+
+          return (
+            <li
+              key={section.id}
+              className={`usa-step-indicator__segment ${liSegmentClassSuffix ? `usa-step-indicator__segment--${liSegmentClassSuffix}` : ""}`}
+              {...(completionState === "current" && { "aria-current": "true" })}
+            >
+              <span className="usa-step-indicator__segment-label">
+                {section.name}
+                {screenreaderStatus && <span className="usa-sr-only">{screenreaderStatus}</span>}
+              </span>
+            </li>
+          );
+        })}
+      </ol>
+
+      {currentSection.shouldHideProgressHeadingAndRequiredMessage !== true && (
+        <div className="usa-step-indicator__header display-flex flex-justify">
+          <h1 className="font-heading-lg">
+            {currentStep !== undefined && (
+              <span className="usa-step-indicator__heading-counter">
+                <span className="usa-sr-only" data-testid="step-text">
+                  Step
+                </span>
+                <span className="usa-step-indicator__current-step">{currentStep}</span>
+                &nbsp;
+                <span className="usa-step-indicator__total-steps">{`of ${currentSection.numSteps}`}</span>
+                &nbsp;
+              </span>
+            )}
+            <span className="usa-step-indicator__heading-text">{currentSection.name}</span>
+          </h1>
+          <div className="text-right">
+            A red asterisk (<RequiredMarker />) indicates a required field.
+          </div>
+        </div>
+      )}
+    </div>
+  );
+};
