+++ src/app/form/page.tsx
@@ -2,7 +2,7 @@
 
 import dynamic from "next/dynamic";
 
-const ClientRoutes = dynamic(() => import("@/app/clientRoutes"), { ssr: false });
+const ClientRoutes = dynamic(() => import("@/app/form/clientRoutes"), { ssr: false });
 
 export default function Page() {
   return <ClientRoutes />;
